kotlin 3.7k

+++ src/main/java/calebzhou/rdi/microservice/model/json/weather/Cloudrate.kt
@@ -7,7 +7,7 @@ data class Cloudrate (
 
   @SerializedName("date" ) var date : String? = null,
   @SerializedName("max"  ) var max  : Double? = null,
-  @SerializedName("min"  ) var min  : Int?    = null,
+  @SerializedName("min"  ) var min  : Double?    = null,
   @SerializedName("avg"  ) var avg  : Double? = null
 
 )
